made lots of progress on airlines
- .trim() string method to bypass accidental spaces on inputs.
- Flight class setters now have the validation process
- Built the admin menu.

I still need to figure out if I need the validation to be on main or just have it on the setters themselves.

If it's on the setters, I can't provide accurate invalid inputs on main without creating an extra boolean variable with it's own setter and getter.

If it's on main, I have to pass the inputs through the valid object first, but somehow I need to account for letter inputs in the double flightPrice setter.
I would need a temporary string variable to hold the input and put it on a try catch test, which would be annoying.

I'll figure out everything at some point